FDG-7813 Apply revenue trends chart changes
